Update README.md
HOMEWORK 1 Q 3 
a. The purpose of these programs is to calculate data and analyze it. 
b. For Q1, the input is the tendered and cost. For Q2, the input is the initial temperature before being converted. 
c. For Q1, the output is the amount of change needed to be given, and the best way to give those bills and coins to someone. For Q2, the output is the new temperature value in Celsius, Fahrenheit, or Kelvin.
d. The execution is debugging or the template. 
e. My program could be improved for Q1, because it is a very inefficient way of calculating the number of bills and coins. If I would have used some sort of function it would have been a lot better and more efficient.